[Refactor] friend 패키지 이름 변경 및 mypage 패키지로 이동
friend 패키지는 mypage에서만 사용하기에 패키지 이동

친구 검색 기능 관련 패키지로 friend > friendsearch로 이름 변경
